Add product images and update showProduct.blade.php
This commit includes the addition of product images and an update to the showProduct.blade.php file. Each image represents a product and is stored in the 'public/product_images' directory. The 'showProduct.blade.php' file has been modified to display these images along with relevant product details.

Images added:
- 0a44238c-9ad9-49de-b1c6-f713a3391d26-th (12).jpeg
- 254e5ce1-0b7a-4419-96fc-a9cade70bbd1-th (6).jpeg
- bc5def65-b58a-48da-b089-6a4876068f12-th (2).jpeg
- dde401ca-2886-40bf-b476-cf155afaa5c7-th (8).jpeg
- f899fe70-6af0-4f2e-942d-46422d6719f1-th (9).jpeg

Modified file:
- resources/views/Products/showProduct.blade.php

These changes enhance the visual representation of products on the website